Level-0: Implement chatbot greeting and exit functionality
